make Parse verify the chain of proxies
This commit changes Parse so that it verifies the chain of proxies
correctly. First, we should not skip any invalid proxies, this is the
equivalent of ignoring errors.

Secondly, we need to iterate from the end of the xff list towards the
front, as that is what the most common (all?) proxies do way. We also
need to verify each proxy is in the allowed list and stop when a failure
occurs since we can no longer trust further values.